Refactored exception handling.
