+++ EyeCT4Rails/EyeCT4RailsLogic/Exceptions/CustomException.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EyeCT4RailsLogic.Exceptions
+{
+    public class CustomException : Exception
+    {
+        public CustomException(string message) : base(message)
+        {
+            
+        }
+
+        public CustomException(string message, Exception innerException) : base(message, innerException)
+        {
+            
+        }
+    }
+}
