+++ EyeCT4Rails/EyeCT4RailsLogic/Exceptions/InvalidDateException.cs
@@ -2,7 +2,7 @@
 
 namespace EyeCT4RailsLogic.Exceptions
 {
-    public class InvalidDateException : Exception
+    public class InvalidDateException : CustomException
     {
         public InvalidDateException(string message) : base(message)
         {
