+++ EyeCT4Rails/EyeCT4RailsLogic/Exceptions/InvalidIdException.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EyeCT4RailsLogic.Exceptions
+{
+    public class InvalidIdException : CustomException
+    {
+        public InvalidIdException(string message) : base(message)
+        {
+            
+        }
+    }
+}
